feat: Moved currency, symbol and kline to app as they are repeated often.
