+++ components/bingx/actions/market-interface/bingx-get-kline-data.mjs
@@ -1,7 +1,5 @@
 import bingx from "../../bingx.app.mjs";
-import {
-  KLINE_DESC_LIST, KLINE_DESC_MAPPING,
-} from "../../bingx-common.mjs";
+import { KLINE_DESC_MAPPING } from "../../bingx-common.mjs";
 
 export default {
   name: "BingX Market GetLatestKline",
@@ -12,23 +10,16 @@ export default {
   props: {
     bingx,
     symbol: {
-      label: "Symbol",
-      description: "Symbol/Ticker/Trading Pair. " +
-                "There must be a hyphen/ \"-\" in the trading pair symbol. eg: BTC-USDT",
-      type: "string",
-      optional: false,
-      default: "BTC-USDT",
-      async options() {
-        const contractsData = await this.bingx.getAllMarketContracts();
-        return contractsData.data.contracts.map((contract) => contract.symbol);
-      },
+      propDefinition: [
+        bingx,
+        "symbol",
+      ],
     },
     klineType: {
-      label: "K-Line Type",
-      description: "The type of K-Line (minutes, hours, weeks etc.)",
-      type: "string",
-      options: KLINE_DESC_LIST,
-      optional: false,
+      propDefinition: [
+        bingx,
+        "klineType",
+      ],
     },
   },
   type: "action",
